refactor typeScript configuration and build script
Removed the 'cross-env-shell' and 'BABEL_ENV=build' from the build command in the script section of package.json. The TypeScript configurations were modified by deleting the global tsconfig.json and adding the extension '@rhino-ui/ts-config/base.json' in the local tsconfig.json. The '@rhino-ui/ts-config' was also added as a workspace requirement in devDependencies.